增加用户只管理自己内容的功能,增加站点配置的数据权限

+++ publiccms-parent/publiccms-core/src/main/java/com/publiccms/controller/admin/cms/CmsPlaceAdminController.java
@@ -21,12 +21,17 @@ import com.publiccms.common.tools.ExtendUtils;
 import com.publiccms.common.tools.RequestUtils;
 import com.publiccms.entities.cms.CmsPlace;
 import com.publiccms.entities.log.LogOperate;
+import com.publiccms.entities.sys.SysDept;
+import com.publiccms.entities.sys.SysDeptPageId;
 import com.publiccms.entities.sys.SysSite;
+import com.publiccms.entities.sys.SysUser;
 import com.publiccms.logic.component.template.MetadataComponent;
 import com.publiccms.logic.component.template.TemplateComponent;
 import com.publiccms.logic.service.cms.CmsPlaceAttributeService;
 import com.publiccms.logic.service.cms.CmsPlaceService;
 import com.publiccms.logic.service.log.LogLoginService;
+import com.publiccms.logic.service.sys.SysDeptPageService;
+import com.publiccms.logic.service.sys.SysDeptService;
 import com.publiccms.views.pojo.model.ExtendDataParameters;
 
 /**
@@ -43,6 +48,10 @@ public class CmsPlaceAdminController extends AbstractController {
     private CmsPlaceAttributeService attributeService;
     @Autowired
     private MetadataComponent metadataComponent;
+    @Autowired
+    private SysDeptPageService sysDeptPageService;
+    @Autowired
+    private SysDeptService sysDeptService;
 
     private String[] ignoreProperties = new String[] { "id", "siteId", "status", "userId", "type", "clicks", "path", "createDate",
             "disabled" };
@@ -68,7 +77,16 @@ public class CmsPlaceAdminController extends AbstractController {
             }
             entity.setPath(entity.getPath().replace("//", CommonConstants.SEPARATOR));
             SysSite site = getSite(request);
-            Long userId = ControllerUtils.getAdminFromSession(session).getId();
+            SysUser user = ControllerUtils.getAdminFromSession(session);
+            SysDept dept = sysDeptService.getEntity(user.getDeptId());
+            if (ControllerUtils.verifyNotEmpty("deptId", user.getDeptId(), model)
+                    || ControllerUtils.verifyNotEmpty("deptId", dept, model)
+                    || ControllerUtils.verifyCustom("noright",
+                            !(dept.isOwnsAllPage() || null != sysDeptPageService.getEntity(new SysDeptPageId(user.getDeptId(),
+                                    CommonConstants.SEPARATOR + TemplateComponent.INCLUDE_DIRECTORY + entity.getPath()))),
+                            model)) {
+                return CommonConstants.TEMPLATE_ERROR;
+            }
             if (CommonUtils.empty(entity.getItemType()) || CommonUtils.empty(entity.getItemId())) {
                 entity.setItemType(CmsPlaceService.ITEM_TYPE_CUSTOM);
                 entity.setItemId(null);
@@ -80,16 +98,16 @@ public class CmsPlaceAdminController extends AbstractController {
                 }
                 entity = service.update(entity.getId(), entity, ignoreProperties);
                 if (null != entity) {
-                    logOperateService.save(new LogOperate(site.getId(), userId, LogLoginService.CHANNEL_WEB_MANAGER,
+                    logOperateService.save(new LogOperate(site.getId(), user.getId(), LogLoginService.CHANNEL_WEB_MANAGER,
                             "update.place", RequestUtils.getIpAddress(request), CommonUtils.getDate(), entity.getPath()));
                 }
             } else {
-                entity.setUserId(userId);
+                entity.setUserId(user.getId());
                 entity.setSiteId(site.getId());
                 entity.setStatus(CmsPlaceService.STATUS_NORMAL);
                 service.save(entity);
-                logOperateService.save(new LogOperate(site.getId(), userId, LogLoginService.CHANNEL_WEB_MANAGER, "save.place",
-                        RequestUtils.getIpAddress(request), CommonUtils.getDate(), entity.getPath()));
+                logOperateService.save(new LogOperate(site.getId(), user.getId(), LogLoginService.CHANNEL_WEB_MANAGER,
+                        "save.place", RequestUtils.getIpAddress(request), CommonUtils.getDate(), entity.getPath()));
             }
             String filePath = siteComponent.getWebTemplateFilePath(site, TemplateComponent.INCLUDE_DIRECTORY + entity.getPath());
             Map<String, String> map = ExtendUtils.getExtentDataMap(extendDataParameters.getExtendDataList(),
@@ -101,6 +119,7 @@ public class CmsPlaceAdminController extends AbstractController {
     }
 
     /**
+     * @param path
      * @param ids
      * @param _csrf
      * @param request
@@ -109,13 +128,24 @@ public class CmsPlaceAdminController extends AbstractController {
      * @return view name
      */
     @RequestMapping("refresh")
-    public String refresh(Long[] ids, String _csrf, HttpServletRequest request, HttpSession session, ModelMap model) {
+    public String refresh(String path, Long[] ids, String _csrf, HttpServletRequest request, HttpSession session,
+            ModelMap model) {
         if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
+        SysUser user = ControllerUtils.getAdminFromSession(session);
+        SysDept dept = sysDeptService.getEntity(user.getDeptId());
+        if (ControllerUtils.verifyNotEmpty("deptId", user.getDeptId(), model)
+                || ControllerUtils.verifyNotEmpty("deptId", dept, model)
+                || ControllerUtils.verifyCustom("noright",
+                        !(dept.isOwnsAllPage() || null != sysDeptPageService.getEntity(new SysDeptPageId(user.getDeptId(),
+                                CommonConstants.SEPARATOR + TemplateComponent.INCLUDE_DIRECTORY + path))),
+                        model)) {
             return CommonConstants.TEMPLATE_ERROR;
         }
         if (CommonUtils.notEmpty(ids)) {
             SysSite site = getSite(request);
-            service.refresh(site.getId(), ids);
+            service.refresh(site.getId(), ids, path);
             logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "refresh.place", RequestUtils.getIpAddress(request),
                     CommonUtils.getDate(), StringUtils.join(ids, ',')));
@@ -124,6 +154,7 @@ public class CmsPlaceAdminController extends AbstractController {
     }
 
     /**
+     * @param path
      * @param ids
      * @param _csrf
      * @param request
@@ -132,13 +163,23 @@ public class CmsPlaceAdminController extends AbstractController {
      * @return view name
      */
     @RequestMapping("check")
-    public String check(Long[] ids, String _csrf, HttpServletRequest request, HttpSession session, ModelMap model) {
+    public String check(String path, Long[] ids, String _csrf, HttpServletRequest request, HttpSession session, ModelMap model) {
         if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
+        SysUser user = ControllerUtils.getAdminFromSession(session);
+        SysDept dept = sysDeptService.getEntity(user.getDeptId());
+        if (ControllerUtils.verifyNotEmpty("deptId", user.getDeptId(), model)
+                || ControllerUtils.verifyNotEmpty("deptId", dept, model)
+                || ControllerUtils.verifyCustom("noright",
+                        !(dept.isOwnsAllPage() || null != sysDeptPageService.getEntity(new SysDeptPageId(user.getDeptId(),
+                                CommonConstants.SEPARATOR + TemplateComponent.INCLUDE_DIRECTORY + path))),
+                        model)) {
             return CommonConstants.TEMPLATE_ERROR;
         }
         if (CommonUtils.notEmpty(ids)) {
             SysSite site = getSite(request);
-            service.check(site.getId(), ids);
+            service.check(site.getId(), ids, path);
             logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "check.place", RequestUtils.getIpAddress(request), CommonUtils.getDate(),
                     StringUtils.join(ids, ',')));
@@ -159,12 +200,21 @@ public class CmsPlaceAdminController extends AbstractController {
         if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
             return CommonConstants.TEMPLATE_ERROR;
         }
+        SysUser user = ControllerUtils.getAdminFromSession(session);
+        SysDept dept = sysDeptService.getEntity(user.getDeptId());
+        if (ControllerUtils.verifyNotEmpty("deptId", user.getDeptId(), model)
+                || ControllerUtils.verifyNotEmpty("deptId", dept, model)
+                || ControllerUtils.verifyCustom("noright",
+                        !(dept.isOwnsAllPage() || null != sysDeptPageService.getEntity(new SysDeptPageId(user.getDeptId(),
+                                CommonConstants.SEPARATOR + TemplateComponent.INCLUDE_DIRECTORY + path))),
+                        model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         if (CommonUtils.notEmpty(path)) {
             SysSite site = getSite(request);
             service.delete(site.getId(), path);
-            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
-                    LogLoginService.CHANNEL_WEB_MANAGER, "clear.place", RequestUtils.getIpAddress(request), CommonUtils.getDate(),
-                    path));
+            logOperateService.save(new LogOperate(site.getId(), user.getId(), LogLoginService.CHANNEL_WEB_MANAGER, "clear.place",
+                    RequestUtils.getIpAddress(request), CommonUtils.getDate(), path));
         }
         return CommonConstants.TEMPLATE_DONE;
     }
@@ -178,13 +228,23 @@ public class CmsPlaceAdminController extends AbstractController {
      * @return view name
      */
     @RequestMapping("delete")
-    public String delete(Long[] ids, String _csrf, HttpServletRequest request, HttpSession session, ModelMap model) {
+    public String delete(String path, Long[] ids, String _csrf, HttpServletRequest request, HttpSession session, ModelMap model) {
         if (ControllerUtils.verifyNotEquals("_csrf", ControllerUtils.getAdminToken(request), _csrf, model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
+        SysUser user = ControllerUtils.getAdminFromSession(session);
+        SysDept dept = sysDeptService.getEntity(user.getDeptId());
+        if (ControllerUtils.verifyNotEmpty("deptId", user.getDeptId(), model)
+                || ControllerUtils.verifyNotEmpty("deptId", dept, model)
+                || ControllerUtils.verifyCustom("noright",
+                        !(dept.isOwnsAllPage() || null != sysDeptPageService.getEntity(new SysDeptPageId(user.getDeptId(),
+                                CommonConstants.SEPARATOR + TemplateComponent.INCLUDE_DIRECTORY + path))),
+                        model)) {
             return CommonConstants.TEMPLATE_ERROR;
         }
         if (CommonUtils.notEmpty(ids)) {
             SysSite site = getSite(request);
-            service.delete(site.getId(), ids);
+            service.delete(site.getId(), ids, path);
             logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
                     LogLoginService.CHANNEL_WEB_MANAGER, "delete.place", RequestUtils.getIpAddress(request),
                     CommonUtils.getDate(), StringUtils.join(ids, ',')));
